Changes in directory structure. Files renamed.
