+++ tests/vector_test.cpp
@@ -1,4 +1,4 @@
-#include "vector_test.h"
+#include "vector.h"
 #include <random>
 #include <functional>
 #include <iostream>
